Refactor modules of minikalman_traits
