+++ crates/minikalman-traits/src/kalman.rs
@@ -1,0 +1,3 @@
+mod matrix_types;
+
+pub use matrix_types::*;
