+++ crates/minikalman-traits/src/matrix/data_type.rs
