+++ crates/minikalman/src/buffer_types/kalman_gain_matrix_buffer.rs
@@ -1,9 +1,9 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
+use minikalman_traits::kalman::KalmanGainMatrix;
 
-use crate::type_traits::KalmanGainMatrix;
-use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
-use minikalman_traits::{Matrix, MatrixMut};
+use minikalman_traits::matrix::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
+use minikalman_traits::matrix::{Matrix, MatrixMut};
 
 pub struct KalmanGainMatrixBuffer<const STATES: usize, const MEASUREMENTS: usize, T, M>(
     M,
