+++ crates/minikalman/src/buffer_types/measurement_process_noise_covariance_matrix_buffer.rs
@@ -1,8 +1,8 @@
-use crate::type_traits::MeasurementProcessNoiseCovarianceMatrix;
-use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
-use minikalman_traits::{Matrix, MatrixMut};
+use minikalman_traits::kalman::MeasurementProcessNoiseCovarianceMatrix;
+use minikalman_traits::matrix::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
+use minikalman_traits::matrix::{Matrix, MatrixMut};
 
 pub struct MeasurementProcessNoiseCovarianceMatrixBuffer<const MEASUREMENT: usize, T, M>(
     M,
