+++ crates/minikalman/src/buffer_types/measurement_transformation_matrix_buffer.rs
@@ -1,9 +1,13 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
+use minikalman_traits::kalman::{
+    MeasurementTransformationMatrix, MeasurementTransformationMatrixMut,
+};
 
-use crate::type_traits::{MeasurementTransformationMatrix, MeasurementTransformationMatrixMut};
-use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned, MatrixDataRef};
-use minikalman_traits::{Matrix, MatrixMut};
+use minikalman_traits::matrix::{
+    IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned, MatrixDataRef,
+};
+use minikalman_traits::matrix::{Matrix, MatrixMut};
 
 pub struct MeasurementTransformationMatrixBuffer<
     const MEASUREMENTS: usize,
