+++ crates/minikalman/src/buffer_types/state_vector_buffer.rs
@@ -1,8 +1,8 @@
-use crate::type_traits::StateVector;
-use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
-use minikalman_traits::{Matrix, MatrixMut};
+use minikalman_traits::kalman::StateVector;
+use minikalman_traits::matrix::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
+use minikalman_traits::matrix::{Matrix, MatrixMut};
 
 pub struct StateVectorBuffer<const STATES: usize, T, M>(M, PhantomData<T>)
 where
