+++ crates/minikalman/src/buffer_types/system_matrix_buffer.rs
@@ -1,9 +1,12 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
+use minikalman_traits::kalman::{SystemMatrix, SystemMatrixMut};
 
-use crate::type_traits::{SystemMatrix, SystemMatrixMut};
-use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned, MatrixDataRef};
-use minikalman_traits::{Matrix, MatrixMut};
+use minikalman_traits::matrix::{
+    IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned, MatrixDataRef,
+};
+
+use minikalman_traits::matrix::{Matrix, MatrixMut};
 
 pub struct SystemMatrixBuffer<const STATES: usize, T, M>(M, PhantomData<T>)
 where
