+++ crates/minikalman/src/lib.rs
@@ -37,18 +37,14 @@ mod buffer_builder;
 pub mod buffer_types;
 mod kalman;
 mod kalman_filter_trait;
-mod matrix_types;
 mod measurement;
 mod static_macros;
-mod type_traits;
 
 #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 #[cfg(feature = "alloc")]
 pub use crate::buffer_builder::BufferBuilder;
 pub use crate::kalman::{Kalman, KalmanBuilder};
-pub use crate::matrix_types::*;
 pub use crate::measurement::{Measurement, MeasurementBuilder};
-pub use minikalman_traits as traits;
 
 /// Re-export `num_traits`.
 pub use num_traits;
@@ -60,10 +56,7 @@ pub mod prelude {
     pub use crate::buffer_builder::BufferBuilder;
     pub use crate::buffer_types::*;
     pub use crate::kalman::{Kalman, KalmanBuilder};
-    pub use crate::matrix_types::{IntoInnerData, MatrixData};
     pub use crate::measurement::{Measurement, MeasurementBuilder};
-    pub use crate::traits::*;
-    pub use crate::type_traits::*;
 
     pub use crate::{
         impl_buffer_A, impl_buffer_B, impl_buffer_H, impl_buffer_K, impl_buffer_P, impl_buffer_Q,
